% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{checkbox}
\alias{checkbox}
\title{Convenient wrappers for shiny widgets using the googleForm lingo}
\usage{
checkbox(id, choices)
}
\arguments{
\item{id}{an inputId}

\item{choices}{a list}
}
\description{
Convenient wrappers for shiny widgets using the googleForm lingo
}
