% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{formQ}
\alias{formQ}
\title{Generates UI}
\usage{
formQ(question)
}
\arguments{
\item{question}{A list containing id, type, required, and (optionally) choices}
}
\description{
This is the 'muscle' behind the UI building and taken from the UI of shinyforms. It takes a list(id, type, etc.) and decides what to make. Used with lapply in main googleform() function over questions arguments.
}
