% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{paragraph}
\alias{paragraph}
\title{Convenient wrappers for shiny widgets using the googleForm lingo}
\usage{
paragraph(id)
}
\arguments{
\item{id}{an inputId}
}
\description{
Convenient wrappers for shiny widgets using the googleForm lingo
}
