% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{selectHeight}
\alias{selectHeight}
\title{Convenient wrapper for making a selectInput about choosing a person's height in feet and inches}
\usage{
selectHeight(id)
}
\arguments{
\item{id}{an inputId}
}
\description{
Convenient wrapper for making a selectInput about choosing a person's height in feet and inches
}
