% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{theme_quickform}
\alias{theme_quickform}
\title{Limited theme option to change color of form}
\usage{
theme_quickform(color)
}
\arguments{
\item{color}{a character string of a hex color code}
}
\description{
Limited theme option to change color of form
}
